医疗健康智能体/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
sentence-transformers
faiss-cpu
tiktoken
orjson
